requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
bcrypt==3.2.0
python-dotenv==0.19.0
Werkzeug==2.2.2
gunicorn==20.1.0
